sync formats

+++ runtime/src/types.rs
@@ -8,7 +8,7 @@ use sp_runtime::{
 };
 
 /// An index to a block.
-pub type BlockNumber = u32;
+pub type BlockNumber = u64;
 
 /// Alias to 512-bit hash when used in the context of a transaction signature on the chain.
 pub type Signature = MultiSignature;
@@ -25,7 +25,7 @@ pub type AccountIndex = u32;
 pub type Amount = i128;
 
 /// Index of a transaction in the chain.
-pub type Index = u32;
+pub type Index = u64;
 
 /// A hash of some data used by the chain.
 pub type Hash = sp_core::H256;
